Add support for creating wgpu surface on macOS (#65)
* Add support for creating wgpu surface on macOS

* Black all the things

* Fix linting

* Add cffi to the dev-requirements as well

This makes it possible to get a working environment when using `pip
install -r dev-requirements.txt`.

* Add comment with some info on Tk on macOS

* Remove cffi from dev-requirements.txt

I do think it is best to keep the requirements as much as possible in
one place, so I'll revert my previous decision to put it in the dev
requirements.
As in: I should have just run: python setup.py develop ;)

* Change Exception type to RuntimeError

This better represents the error. Also adds a message and some comments
to help people who might encounter the error.