Update README.md "OpenGL is...showings its cracks" -> it's/its (#560)
Let's focus on understanding the difference between it's (means *exactly* "it is": the apostrophe in English elides words together, e.g., "don't" = "do not", 'n' = "and"). "Its" is the pronoun for the thing described (OpenGL), so companions to "its" also don't have apostrophes (his, her, his, hers, their). We're programmers: this syntax should not be hard for programmers to commit to knowledge. Even native English speakesr make this error far too often.