Add patches for conda compatibility
For the conda-forge packages I package wgpu-native myself and install
them in the "system" libraries.

This should continue to allow you to use the "wheel" vendored headers
but then also look for the headers in the "system" locations.

I've been dragging on a simpler (less upstreamable) version of this
patch for about a year but I'm hoping I don't have to with this
addition.